Split learningpath embed into multiple files for readability
