+++ src/messages/messagesEN.ts
@@ -344,6 +344,9 @@ const messages = {
     accordionTitle: "Learning path content",
     learningsteps: "Steps",
     stepCompleted: "Completed",
+    externalWarning:
+      "This learning path has been developed by an external teacher, who holds the editorial responsibility. Please note that it may contain texts and links that do not originate from ndla.no.",
+    externalLink: "Open in a new window",
   },
   movedResourcePage: {
     title: "The page has moved, but you can find it here:",
